Move functinality from index to createNodeMenu function
