+++ src/Menu/NodeMenu/NodeMenu.test.ts
@@ -1,6 +1,6 @@
+import MenuItemsCollection from '../MenuItemsCollection/MenuItemsCollection.js';
 import NodeMenuItem from '../NodeMenuItem/NodeMenuItem';
-import { OptionsCollection } from '../OptionsCollection/OptionsCollection.js';
-import { type MenuItem } from '../interfaces/MenuItem';
+import ReadLineNode from '../ReadLineNode/ReadLineNode';
 import NodeMenu from './NodeMenu';
 
 describe('Given a NodeMenu', () => {
@@ -12,10 +12,11 @@ describe('Given a NodeMenu', () => {
 
   describe('when showMenu method is called', () => {
     test('then it should print the menu items collection', () => {
-      const menuOptions = new OptionsCollection<MenuItem>();
+      const readline = new ReadLineNode();
+      const menuOptions = new MenuItemsCollection();
       const option1 = new NodeMenuItem('start', '1');
       menuOptions.add(option1);
-      const nodeMenu = new NodeMenu(menuOptions);
+      const nodeMenu = new NodeMenu(menuOptions, readline);
       nodeMenu.showMenu();
     });
   });
